net/mock: Provide SocketService instead of PeerService
While implementing `PeerService` would've made the peer code slightly
cleaner, it restricted the `NetworkService::Socket` type with too
many trait bounds and ultimately made interface far too inflexible
for different implementations.

This commit fixes that issue by implementing a wrapper type on top
of `TcpStream` and modifies the associated type of `NetworkService`
to only implement `SocketService`. This gives great flexibility
for the network service provider to implement whatever it wants
under the hood while staying compatible with the rest of the networking
code.