p2p: Add empty event handlers for peer
This commit provides the overall design of the event loop of each
individual peer. It contains a function for incoming events, which
dispatch the event to the network manager, a function for events
originating from the network manager which may contain an outbound
messages, and a function for handling periodic and one-shot events
executed when a timer runs out.

The idea is to implement the entire peer-related protocol into
these functions, making it very easy to test and maintain the
entire protocol.