p2p: Do not match multiple variables in one expression
Instead of matching both state and message type at the same time,
match the state first, then call an appropriate function for that
state and match the message type in that function to handle the
(state, message type) combination.

Additionally, add some comments, reword an error message and rename
a few tests functions to be more descriptive.