net/libp2p: Start listening on the specified address
Once the libp2p backend has been spawned, send it a message that
commands it to start listening on the specified address and along
with that command, a sender-half of a oneshot channel.

When the listen_on() command returns, send the result of that through
the oneshot channel back to the libp2p front-end which then processes
the result accordingly and either proceeds with initialization or
exits with error.