p2p: Introduce SwarmManager object
This object is responsible for handling all connecivity-related work
such as auto-connection, responding to incoming connections, handshaking,
and dealing with peer discovery events.

All of that will be transparent to Sync/P2P and they only receive
PeerConnected/PeerDisconnected events (if they are listening to those)
and respond accordingly but they don't, for example, receiving
information if during auto-connection an IPv6 addres of the peer failed
and next up is trying the IPv4 address as that has nothing do with syncing.

In short, SwarmManager allows clean and isolated implementation of
connectivity management.