net/libp2p: Redesign libp2p backend
The idea is to implement the protocol inside event-specific `inject_event()`
handlers and store other auxilliary objects into the `NetworkBehaviour` so
that the protocol handlers will have access to them. This allows the
protocol to be implemented fully within libp2p without having to take
events out of the swarm, process them using the `Backend` object and then
injecting them back to the swarm.

It makes the design cleaner when there's the front-end, the backend
(`Libp2pBehaviour`) and then the command interface which for now will be
called the backend.