common: Separate out `Block` and `Genesis` types
Introduce a type-level distinction between `Genesis` and proper `Block`.
There is an additional abstraction called `GenBlock` representing
something that can be either a proper `Block` or `Genesis`.

This is analogous to a linked list:
 * List roughly corresponds to GenBlock
 * Nil corresponds to Genesis
 * Cons cell roughly corresponds to a Block

Or in a pseudocode:

```
List     = Nil     | Cons  { head: Item,      tail: List       }
GenBlock = Genesis | Block { data: BlockData, parent: GenBlock }
```

Note: In the code, the parent link is actually part of the block (block
header to be more precise), but the high-level structure still holds.

The idea is to be more precise about what could be expected where, and
communicate the information in types:

* If only regular block or only genesis is possible in particular case,
  the type reflects that, and the code does not need to be concerned
  with the "other" case.
* If either is possible, the type is one of `GenBlock` and the code is
  required to handle both cases by using appropriate abstraction or by
  explicitly matching and handling both possibilities.

This is still somewhat rough around the edges and subsequent cleanups
are expected. These include adding some helper methods to clean up
repeating patterns, adding extra conversion and comparison methods, etc.
These would hopefully be either more self-contained or mechanical
changes. So hopefully less disruptive.