net/mock: Implement peer information exchange
This commits provides the mock interface implementation for libp2p's
identify protocol where peers exchange their information before
the connection state is reported to the front end.

The protocol is configured to immediately reject peers that are in
different networks and all other information is relayed to `PeerManager`
which can then decided whether the node version and set of provided
protocols are enough for maintaining the communication with the node.